moving the array of users and the method of getting posts to the store
